requirements.txt
PyMuPDF==1.23.5
icecream==2.1.3
pdfminer.six==20221105

